requirements.txt
# python 3.10.0
ipykernel
pandas
numpy
tensorflow-directml-plugin
streamlit==1.13.0
flask
google-api-python-client
python-dotenv